feat(cloud-backup): Requires copy_policy_items_enabled for both copy request flags and simplifies changes
